requirements.txt
tensorflow-gpu==1.14.0	# 理论上任何1.1*版本的都可以
numpy==1.18.1		# 理论上任何版本的都可以
opencv-python==4.1.2.30	# 理论上任何版本的都可以
